Incorporate regression containers and reorg forecast containers
