+++ src/covid_model_seiir_pipeline/pipeline/forecasting/model/__init__.py
@@ -1,3 +1,12 @@
+from covid_model_seiir_pipeline.pipeline.forecasting.model.containers import (
+    HospitalFatalityRatioData,
+    HospitalCensusData,
+    HospitalMetrics,
+    HospitalCorrectionFactors,
+    CompartmentInfo,
+    ScenarioData,
+    OutputMetrics,
+)
 from covid_model_seiir_pipeline.pipeline.forecasting.model.ode_forecast import (
     run_normal_ode_model_by_location,
     forecast_beta,
@@ -15,4 +24,8 @@ from covid_model_seiir_pipeline.pipeline.forecasting.model.mandate_reimposition 
     compute_mobility_lower_bound,
     compute_new_mobility,
     unpack_parameters
+)
+# Just want to expose from this namespace
+from covid_model_seiir_pipeline.pipeline.regression.model import (
+    get_death_weights,
 )
