+++ src/covid_model_seiir_pipeline/pipeline/regression/__init__.py
@@ -5,7 +5,6 @@ from covid_model_seiir_pipeline.pipeline.regression.specification import (
 )
 from covid_model_seiir_pipeline.pipeline.regression.data import (
     RegressionDataInterface,
-    HospitalFatalityRatioData,
 )
 from covid_model_seiir_pipeline.pipeline.regression.task import (
     beta_regression,
