+++ src/covid_model_seiir_pipeline/pipeline/regression/model/__init__.py
@@ -18,6 +18,4 @@ from covid_model_seiir_pipeline.pipeline.regression.model.hospital_corrections i
     get_death_weights,
     compute_hospital_usage,
     calculate_hospital_correction_factors,
-    HospitalMetrics,
-    HospitalCorrectionFactors,
 )
